Rework the tests.
Each time I make changes here, it's fairly difficult to write tests,
because the tests tend to use weird mocking, but also use Rabbit
internally.

Originally, I wanted to add one line of code and could not cover it with
a test, given how things were set up.

So, this reimagines the test suite as Integration Tests and Unit Tests.

Integration Tests
=================

Since most of what this library does is wrap Bunny in an abstracation,
it seems logical to cover most of this libray with tests that
actually use Rabbit.  The legacy tests kind did this, so I started
with the tests in spec/integration.

These require that Rabbit and Redis be running, and they test the normal
happy path stuff, simulating how someone would use this library to send
and receive messages.  These specs along cover quite a bit of this code.

Tests here should never ever mock.

Unit Tests
==========

For odder stuff, there are unit tests.  These use mocking and do not
require Rabbit to be running at all.  In fact, these tests should
never require Rabbit.

These tests cover stuff that's too cumbersome to cover in an integration
test or that are easier covered with typicaly unit-testing practices.

Legacy Tests
============

These are in spec/legacy and are the original tests.  Some of them were
moved to spec/unit and very slightly modified.  The rest remain and are
not run as part of the normal test suite.  They are now redundant, as
far as I'm concerned, but I've left them here just to be safe.

Conclusion
==========

I *think* this makes the overall test suite more sane and easier to
undertand.  It should also create a clear boundary between tests that
need Rabbit and those that don't.

Note that we didn't really touch any of the production code in order to
do this.