Adds integration spec for exceptions in error handlers
We experienced an issue in COGS where an exception in the error_handling_chain
broke the chain unintentionally and left us with a worker that seemed to be
alive but was no longer processing errors.

There was no diagnostic information for us to determine what broke and how.

See https://docs.google.com/document/d/1BqwBP3two_w3m4KuFbR57yQwUoHng6T-XVmPNBmiWM0/edit#heading=h.udep7n7vrmim for the specific post-mortem.

In this commit, I created an integration spec that manually demonstrated that an
exception in an error handler prevents subsequent error handlers from running.