Add support for multiple routing keys
I extended the existing ROUTING_KEY option to handle a list of comma
separated routing keys.  Right now for sku service, we're manually going
throught the webconsole to add more routing keys to the queue.  This is
not only tedious, but creates opportunity for confusion since the value
of the `ROUTING_KEY` passed to the rake task doesn't tell the whole
story of which messages a handler will recieve.

This change is backwards compatible, so should be safe to release as a
minor version bump.