Updates Pwwka::ErrorHandlers::Chain to handle exceptions
and adds unit specs verifying the behavior.

If an error occurs during the error handling chain, we want to log the exception
that occurs in the error handler in the error handling chain and cause the Ruby
process to exit.

One interesting aspect of the implementation is that calling `abort` or `exit`
without wrapping the appropriate line in the spec in an `expect {}` block leads
to, depending upon the order in which the specs are run, an early exit of the
test process and subsequent specs not being run.

For that reason, every spec exercising this functionality requires an `expect
{}` block.