Halt the error chain instead of crashing the Ruby process
Instead of `abort`ing and killing the process, we call false. This implies that
users of this should implement some type of monitoring and alerting system which
alerts on fatal log messages from pwwka to catch incidences of fragile,
unresilient error handlers.

Additionally, the spec has been reduced to the bare minimum of setup needed to
prove its case, the indentation has been fixed, and the spec wording updated.